fix(blocks): restore standalone tab render path
