+++ apps/mesh/src/web/layouts/main-panel-tabs/index.tsx
@@ -14,7 +14,8 @@ import { SettingsTab } from "./settings-tab";
 import { OverviewTab } from "./overview-tab";
 import { TaskBoardPage } from "@/web/layouts/task-board";
 import { GitTab } from "@/web/components/thread/github/git-tab";
-import { PreviewBlocksTab } from "./preview-blocks-tab";
+import { PreviewTab } from "./preview-tab";
+import { BlocksTab } from "./blocks-tab";
 import { CodeTab } from "./code-tab";
 import { ContentTab } from "./content-tab";
 import { AutomationTab } from "./automation-tab";
@@ -92,8 +93,11 @@ function TabBody({
   if (activeTab === "automations") {
     return <AutomationsListTab virtualMcpId={virtualMcpId} />;
   }
-  if (activeTab === "preview" || activeTab === "blocks") {
-    return <PreviewBlocksTab surface={activeTab} virtualMcpId={virtualMcpId} />;
+  if (activeTab === "preview") {
+    return <PreviewTab virtualMcpId={virtualMcpId} />;
+  }
+  if (activeTab === "blocks") {
+    return <BlocksTab virtualMcpId={virtualMcpId} />;
   }
   const codeTab = parseCodeTabId(activeTab);
   if (codeTab) {
@@ -175,16 +179,8 @@ export function MainPanelContent({
       taskId,
     });
 
-  // Preview and Blocks render one shared `PreviewContent` (see PreviewBlocksTab),
-  // so they must sit under the SAME ErrorBoundary instance — keying it per-tab
-  // would remount the subtree and reload the preview iframe on every switch.
-  const boundaryKey =
-    activeTab === "preview" || activeTab === "blocks"
-      ? "preview-blocks"
-      : activeTab;
-
   return (
-    <ErrorBoundary key={boundaryKey}>
+    <ErrorBoundary key={activeTab}>
       <Suspense fallback={<MainPanelLoading />}>
         <TabBody
           activeTab={activeTab}
