+++ apps/mesh/src/web/layouts/main-panel-tabs/source-system-tabs.test.ts
@@ -2,10 +2,10 @@ import { describe, expect, test } from "bun:test";
 import { getSourceSystemTabs } from "./source-system-tabs";
 
 describe("getSourceSystemTabs", () => {
-  test("returns Blocks, Preview, and Code for clonable source", () => {
+  test("returns Preview, Blocks, and Code for clonable source", () => {
     expect(getSourceSystemTabs(true)).toEqual([
-      { id: "blocks", title: "Blocks" },
       { id: "preview", title: "Preview" },
+      { id: "blocks", title: "Blocks" },
       { id: "code", title: "Code" },
     ]);
   });
